+++ data/examples/declaration/splice/bracket-declaration-out.hs
@@ -24,3 +24,12 @@ $( singletons
          deriving (Eq, Ord, Enum, Bounded, Show)
        |]
  )
+
+foo = [d| type X = * |]
+
+foo =
+  [d|
+    type X = *
+
+    data A
+    |]